# Use arrays rather than Dicts for Contour Chasing