This builds on #50. Inspired by: https://www.researchgate.net/publication/282975362_Flying_Edges_A_High-Performance_Scalable_Isocontouring_Algorithm

This could serve as a basis for that implementation, which should yield even better performance once fully implemented.

The basic idea is to store each case in a 2D Matrix and loop through this matrix (once). Since the matrix is all UInt8, it ends up very compact in memory. We also avoid hashing overhead and do not need to store keys. Since we track the number of non-zero elements for large arrays, sparse contours (such as one closed loop) should still perform better than the dictionary approach, though they may require more memory allocations. 

For reference:
v0.5.1:
```
contour
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "testdata" => BenchmarkTools.Trial: 
	  memory estimate:  3.14 MiB
	  allocs estimate:  31197
	  --------------
	  minimum time:     2.951 ms (0.00% GC)
	  median time:      3.283 ms (0.00% GC)
	  mean time:        3.464 ms (3.25% GC)
	  maximum time:     7.878 ms (0.00% GC)
	  --------------
	  samples:          1438
	  evals/sample:     1

```

This PR:
```
contour
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "testdata" => BenchmarkTools.Trial: 
	  memory estimate:  569.73 KiB
	  allocs estimate:  370
	  --------------
	  minimum time:     911.743 μs (0.00% GC)
	  median time:      1.029 ms (0.00% GC)
	  mean time:        1.068 ms (1.40% GC)
	  maximum time:     2.901 ms (58.94% GC)
	  --------------
	  samples:          4662
	  evals/sample:     1
```