@sjkelly 2020-03-02
Here is a case where allocations are greater but performance is better:
```
using Contour
using BenchmarkTools

x = -1:0.0001:1
y = -1:0.0001:1
c(x,y) = sqrt(x^2+y^2) - 1

z = [c(xi,yi) for xi in x, yi in y];

@benchmark contours(collect(x), collect(y), z)
```
Master:
```
BenchmarkTools.Trial: 
  memory estimate:  103.07 MiB
  allocs estimate:  795684
  --------------
  minimum time:     37.911 s (0.03% GC)
  median time:      37.911 s (0.03% GC)
  mean time:        37.911 s (0.03% GC)
  maximum time:     37.911 s (0.03% GC)
  --------------
  samples:          1
  evals/sample:     1
```

PR:
```
BenchmarkTools.Trial: 
  memory estimate:  396.66 MiB
  allocs estimate:  291
  --------------
  minimum time:     12.295 s (0.00% GC)
  median time:      12.295 s (0.00% GC)
  mean time:        12.295 s (0.00% GC)
  maximum time:     12.295 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1
```